+++ app/Http/Controllers/ItemController.php
@@ -478,11 +478,11 @@ class ItemController extends Controller
         }
 
         $categories = (new Category())->paginatedCollection(
+            $resource_type_id,
             $this->permitted_resource_types,
             $this->include_public,
             0,
-            100,
-            ['resource_type'=>$resource_type_id]
+            100
         );
 
         foreach ($categories as $category) {